+++ src/main/java/com/kalsym/order/service/service/WhatsappService.java
@@ -44,6 +44,15 @@ public class WhatsappService {
     @Value("${whatsapp.service.order.reminder.refid:60133429331}")
     private String orderReminderRefId;
     
+    @Value("${whatsapp.service.admin.alert.templatename:welcome_to_symplified_7}")
+    private String adminAlertTemplateName;
+    
+    @Value("${whatsapp.service.admin.alert.refid:60133429331}")
+    private String adminAlertRefId;
+    
+    @Value("${whatsapp.service.admin.msisdn:60133429331}")
+    private String adminMsisdn;
+    
     public boolean sendOrderReminder(String[] recipients, String storeName, String invoiceNo, String orderId, String merchantToken) throws Exception {
         String logprefix = "sendWhatsappMessage";
         RestTemplate restTemplate = new RestTemplate();        
@@ -70,7 +79,40 @@ public class WhatsappService {
             Logger.application.info(Logger.pattern, OrderServiceApplication.VERSION, logprefix, "res: " + res.getBody(), "");
             return true;
         } else {
-            Logger.application.info(Logger.pattern, OrderServiceApplication.VERSION, logprefix, "could not send verification email res: " + res, "");
+            Logger.application.info(Logger.pattern, OrderServiceApplication.VERSION, logprefix, "could not send sendOrderReminder res: " + res, "");
+            return false;
+        }
+
+    }
+    
+    
+    public boolean sendAdminAlert(String storeName, String invoiceNo, String orderId) throws Exception {
+        String logprefix = "sendAdminAlert";
+        String[] recipients = {adminMsisdn};
+        RestTemplate restTemplate = new RestTemplate();        
+        HttpHeaders headers = new HttpHeaders();
+        WhatsappMessage request = new WhatsappMessage();
+        request.setGuest(false);
+        request.setRecipientIds(recipients);
+        request.setRefId(recipients[0]);
+        request.setReferenceId(adminAlertRefId);
+        request.setOrderId(orderId);
+        Template template = new Template();
+        template.setName(adminAlertTemplateName);
+        String[] message = {storeName, invoiceNo};
+        template.setParameters(message);
+        request.setTemplate(template);
+        HttpEntity<WhatsappMessage> httpEntity = new HttpEntity<>(request, headers);
+        Logger.application.info(Logger.pattern, OrderServiceApplication.VERSION, logprefix, "url: " + whatsappServiceUrl, "");
+        Logger.application.info(Logger.pattern, OrderServiceApplication.VERSION, logprefix, "httpEntity: " + httpEntity, "");
+
+        ResponseEntity<String> res = restTemplate.postForEntity(whatsappServiceUrl, httpEntity, String.class);
+
+        if (res.getStatusCode() == HttpStatus.ACCEPTED || res.getStatusCode() == HttpStatus.OK) {
+            Logger.application.info(Logger.pattern, OrderServiceApplication.VERSION, logprefix, "res: " + res.getBody(), "");
+            return true;
+        } else {
+            Logger.application.info(Logger.pattern, OrderServiceApplication.VERSION, logprefix, "could not send sendAdminAlert res: " + res, "");
             return false;
         }